# Conflicts:
#	composer.json
#	src/auth/.travis.yml
#	src/auth/README.md
#	src/auth/composer.json
#	src/auth/phpunit.xml
#	src/auth/src/AuthManager.php
#	src/auth/src/AuthUserService.php
#	src/auth/src/Bean/AuthResult.php
#	src/auth/src/Bean/AuthSession.php
#	src/auth/src/Bootstrap/CoreBean.php
#	src/auth/src/Exception/AuthException.php
#	src/auth/src/Helper/ErrorCode.php
#	src/auth/src/Helper/ErrorCodeHelper.php
#	src/auth/src/Mapping/AccountTypeInterface.php
#	src/auth/src/Mapping/AuthHandlerInterface.php
#	src/auth/src/Mapping/AuthManagerInterface.php
#	src/auth/src/Mapping/AuthServiceInterface.php
#	src/auth/src/Mapping/AuthorizationParserInterface.php
#	src/auth/src/Mapping/TokenParserInterface.php
#	src/auth/src/Middleware/AclMiddleware.php
#	src/auth/src/Middleware/AuthMiddleware.php
#	src/auth/src/Parser/AuthorizationHeaderParser.php
#	src/auth/src/Parser/Handler/BasicAuthHandler.php
#	src/auth/src/Parser/Handler/BearerTokenHandler.php
#	src/auth/src/Parser/JWTTokenParser.php
#	src/auth/test/.env
#	src/auth/test/Cases/Account/TestAccount.php
#	src/auth/test/Cases/Handler/BasicAuthParserTest.php
#	src/auth/test/Cases/Handler/BearerTokenParserTest.php
#	src/auth/test/Cases/Helper/ErrorCodeHelperTest.php
#	src/auth/test/Cases/Parser/JWTTokenParserTest.php
#	src/auth/test/config/beans/base.php
#	src/auth/test/config/properties/cache.php
#	src/bean/src/Resource/AnnotationResource.php
#	src/bean/src/Wrapper/AbstractWrapper.php
#	src/db/.travis.yml
#	src/db/src/Executor.php
#	src/db/src/Statement.php
#	src/db/test/Cases/Mysql/BugTest.php
#	src/db/test/Cases/Mysql/QueryBuilderTest.php
#	src/db/test/Testing/Entity/User2.php
#	src/framework/src/Aop/Aop.php
#	src/framework/src/Bean/Resource/ServerAnnotationResource.php
#	src/framework/src/Bootstrap/Server/AbstractServer.php
#	src/framework/src/Core/RequestContext.php
#	src/framework/src/Helper/PhpHelper.php
#	src/framework/src/Proxy/Proxy.php
#	src/framework/src/Testing/SwooleResponse.php
#	src/framework/test/Cases/BeanTest.php
#	src/http-client/src/HttpCoResult.php
#	src/http-client/test/Cases/CoroutineClientTest.php
#	src/http-client/test/Cases/CurlClientTest.php
#	src/http-server/src/Command/ServerCommand.php
#	src/http-server/src/Router/HandlerAdapter.php
#	src/http-server/src/Validator/HttpValidator.php
#	src/i18n/src/Translator.php
#	src/process/src/ProcessBuilder.php
#	src/redis/.travis.yml
#	src/redis/src/Redis.php
#	src/redis/src/RedisConnection.php
#	src/redis/test/Cases/ServerTest.php
#	src/redis/test/Cases/ZsetTest.php
#	src/rpc-client/composer.json
#	src/rpc-client/src/Service/ServiceProxy.php
#	src/rpc-client/test/config/properties/app.php
#	src/rpc-server/src/Bootstrap/Listeners/RpcEventListener.php
#	src/rpc/src/Packer/EofTrait.php
#	src/rpc/test/Cases/JsonPackerTest.php
#	src/service-governance/src/Circuit/CircuitBreaker.php
#	src/service-governance/src/Circuit/CloseState.php
#	src/service-governance/src/Circuit/HalfOpenState.php
#	src/service-governance/src/Provider/ConsulProvider.php
#	src/session/README.md
#	src/session/src/Handler/FileSessionHandler.php
#	src/session/src/Handler/NeedRequestInterface.php
#	src/session/src/Handler/RedisSessionHandler.php
#	src/task/src/QueueTask.php
#	src/task/src/Task.php
#	src/task/src/TaskExecutor.php

+++ src/framework/src/Pool/PoolConfigInterface.php
@@ -27,9 +27,9 @@ interface PoolConfigInterface extends Arrayable
     public function getMaxWait(): int;
 
     /**
-     * @return int
+     * @return float
      */
-    public function getTimeout(): int;
+    public function getTimeout(): float;
 
     /**
      * @return array
